+++ o3seespy/__init__.py
@@ -16,5 +16,7 @@ from .command import parallelcmds
 
 try:
     from custom_openseespy import opensees as opy
+    from custom_openseespy import opensees as ops
 except ModuleNotFoundError:
     import openseespy.opensees as opy
+    import openseespy.opensees as ops